requirements.txt
streamlit
flask
flask-cors
flasgger
openai
python-dotenv
PyMuPDF
Pillow
